# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForSEII2526.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Compras/GetCompra": {
      "get": {
        "tags": [
          "Compras"
        ],
        "operationId": "GetCompra",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBocadilloForCreateDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBocadilloForCreateDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBocadilloForCreateDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras/CrearCompra": {
      "post": {
        "tags": [
          "Compras"
        ],
        "operationId": "CrearCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraBocadilloForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraBocadilloForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompraBocadilloForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBocadilloDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBocadilloDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraBocadilloDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompraBocadilloDetailDTO": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellido1_cliente": {
            "type": "string",
            "nullable": true
          },
          "apellido2_cliente": {
            "type": "string",
            "nullable": true
          },
          "metodoPagoId": {
            "type": "integer",
            "format": "int32"
          },
          "bocadillosComprados": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraBocadilloItemDTO"
            },
            "nullable": true
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "precio": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "CompraBocadilloForCreateDTO": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "apellido1_cliente": {
            "type": "string",
            "nullable": true
          },
          "apellido2_cliente": {
            "type": "string",
            "nullable": true
          },
          "metodoPagoId": {
            "type": "integer",
            "format": "int32"
          },
          "bocadillosComprados": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraBocadilloItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraBocadilloItemDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "tipoPan": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}